Autoupdate functionality
This provides MBE like autoupdate functionality for the BufferGator
(notably NOT the Tab Buffer Catalog, only the listed buffers version.)
In order to make this a little easier to use, it also sets up "buffer
status symbols" for current, alternate (and modified, which already
existed) and changes the syntax definitions to the accommodate this.
in a quasi-related change, it keeps track of the current buffer and
uses the `set currentline` function to highlight it rather than it's
own function.