Add option to exclusively MRU cycle through the two MRU buffers
In my humble opinion, it's almost impossible to remember more than the
last two used buffers. Therefore, it would make sense that there's an
option to enable cycling through the last two used buffers, which is
what I want most of the time. I am confident I'm not alone in this.

This patch implements that functionality.